Documentation of struggle:
In component User, I did a Link that passed props:

      <Link to={{
        pathname: `/user/${this.props.id}`,
        state: {
          name: this.props.name,
          nick: this.props.nick
        }
      }}>

The reason is that I wanted to fetch the same information only once. So instead of fetching info about the user in the Hero component, I passed the information that was fetched in UserList, as props to User, and in User as props in Link (available in /user/:wildcard).

But, when I now code the callback in App, and send this function as props in the Route, I get a crash:

<Route exact path="/user/:id" render={() => <TaskView nrOfPendingTasks={this.nrOfPendingTasks} />} />

This works when there's no /:wildcard, but doesn't work with it. To keep on going forward in the project I will fetch the data about the users twice; one for the UserList, and one for the Hero.