Bug fix
The parallax function, and the addEventListener, was in the render(). When going back from TaskView/Hero to ListView, the event listener was still on the window, causing the background.style to be used on undefined (because .hero didn't exist) so the app broke.

First I fixed the breaking with adding an if-else; only if the .hero existed, the style.top would be used. But with the console.log in both the if and the else, I could see that the computer was working all the time; because the event listener, the console log was always printing the status of the .hero.

Fixed when using removeEventListener, and put the two methods in comonentDidMount respective componentWillUnmount.

Now instead I get this in the console:
"SCRIPT5007: Unable to get property 'closest' of undefined or null reference"